Allow overriding the QML files used by placing them in /usr/share/signon-ui/qml
If there is a webview.qml file in that directory, load it instead of using 
the one built into the resources.  This is useful for debugging QML only 
changes on the device without needing to rebuild and install signon-ui.

It is important to note, when overriding them you need to copy include the .qml and
.js files into that directory.